Changing the storage function to store only finished or faulty resources
